refactor GroupRepository
