+++ web/backend/communect/src/main/kotlin/com/example/communect/infrastructure/db/mapper/UserGroupMapper.kt
@@ -11,7 +11,6 @@ import com.example.communect.infrastructure.db.mapper.UserGroupDynamicSqlSupport
 import com.example.communect.infrastructure.db.mapper.UserGroupDynamicSqlSupport.userGroup
 import com.example.communect.infrastructure.db.mapper.UserGroupDynamicSqlSupport.usergroupid
 import com.example.communect.infrastructure.db.mapper.UserGroupDynamicSqlSupport.userid
-import com.example.communect.infrastructure.db.record.NoticeGroup
 import com.example.communect.infrastructure.db.record.UserGroup
 import org.apache.ibatis.annotations.*
 import org.apache.ibatis.type.EnumTypeHandler
@@ -19,19 +18,8 @@ import org.apache.ibatis.type.JdbcType
 import org.mybatis.dynamic.sql.insert.render.InsertStatementProvider
 import org.mybatis.dynamic.sql.select.render.SelectStatementProvider
 import org.mybatis.dynamic.sql.util.SqlProviderAdapter
-import org.mybatis.dynamic.sql.util.kotlin.CountCompleter
-import org.mybatis.dynamic.sql.util.kotlin.DeleteCompleter
-import org.mybatis.dynamic.sql.util.kotlin.KotlinUpdateBuilder
-import org.mybatis.dynamic.sql.util.kotlin.SelectCompleter
-import org.mybatis.dynamic.sql.util.kotlin.UpdateCompleter
-import org.mybatis.dynamic.sql.util.kotlin.mybatis3.countFrom
-import org.mybatis.dynamic.sql.util.kotlin.mybatis3.deleteFrom
-import org.mybatis.dynamic.sql.util.kotlin.mybatis3.insert
-import org.mybatis.dynamic.sql.util.kotlin.mybatis3.insertMultiple
-import org.mybatis.dynamic.sql.util.kotlin.mybatis3.selectDistinct
-import org.mybatis.dynamic.sql.util.kotlin.mybatis3.selectList
-import org.mybatis.dynamic.sql.util.kotlin.mybatis3.selectOne
-import org.mybatis.dynamic.sql.util.kotlin.mybatis3.update
+import org.mybatis.dynamic.sql.util.kotlin.*
+import org.mybatis.dynamic.sql.util.kotlin.mybatis3.*
 import org.mybatis.dynamic.sql.util.mybatis3.CommonCountMapper
 import org.mybatis.dynamic.sql.util.mybatis3.CommonDeleteMapper
 import org.mybatis.dynamic.sql.util.mybatis3.CommonInsertMapper
